Refactor CheckoutForm component to use CustomerInfoForm and PaymentForm components
